requirements.txt
streamlit
scikit-learn
pandas
nltk
python-docx
PyMuPDF
reportlab